refactor dashboard code into seperate files and rename as app
