+++ dashboard/README.md
@@ -1,3 +1,3 @@
 # Dashboard
 
-- [`dashboard.py`](dashboard.py): [FastHTML](https://fastht.ml/) dashboard for the project (WIP).
+- [`app.py`](app.py): [FastHTML](https://fastht.ml/) dashboard for the project (WIP).
