+++ dashboard/constants.py
@@ -1,0 +1,2 @@
+DEFAULT_LOAD_N_CHARTS = 10
+DEFAULT_ALERT_MAX_N = 30
