+++ dashboard/routes.py
@@ -1,0 +1,439 @@
+from fasthtml.common import *
+from monsterui.all import *
+from fasthtml.svg import *
+
+from app import app, rt
+from constants import *
+from data import get_data
+from components import _create_controls
+from charts import ChartManager
+
+
+@rt
+def index(request: Request):
+    # Check if this is an HTMX request
+    is_htmx = request.headers.get("HX-Request") == "true"
+
+    # Add dark mode class to body if enabled
+    script = Script(
+        f"""
+        if ({'true' if app.state.dark_mode else 'false'}) {{
+            document.body.classList.add('dark-mode');
+        }}
+    """
+    )
+
+    # Get batch stats
+    batch_stats = {}
+    for batch_name in app.state.metric_batches:
+        if batch_name not in app.state.df_cache:
+            df = get_data(
+                app.state.specs_enabled[batch_name], max_n=DEFAULT_ALERT_MAX_N
+            )
+            app.state.df_cache[batch_name] = df
+        else:
+            df = app.state.df_cache[batch_name]
+
+        latest_timestamp = df["metric_timestamp"].max() if not df.empty else "No data"
+        if latest_timestamp != "No data":
+            from datetime import datetime
+
+            # Parse the ISO format timestamp and format it to show date and time
+            dt = datetime.fromisoformat(latest_timestamp.replace("Z", "+00:00"))
+            latest_timestamp_str = dt.strftime("%Y-%m-%d %H:%M:%S")
+
+            # Calculate time difference
+            now = datetime.now(dt.tzinfo)
+            diff_seconds = (now - dt).total_seconds()
+
+            # Choose appropriate time unit
+            if diff_seconds < 3600:  # Less than 1 hour
+                minutes_ago = round(diff_seconds / 60, 1)
+                time_ago_str = (
+                    f"({minutes_ago:.1f} minute{'s' if minutes_ago != 1 else ''} ago)"
+                )
+            elif diff_seconds < 86400:  # Less than 24 hours
+                hours_ago = round(diff_seconds / 3600, 1)
+                time_ago_str = (
+                    f"({hours_ago:.1f} hour{'s' if hours_ago != 1 else ''} ago)"
+                )
+            else:  # Days or more
+                days_ago = round(diff_seconds / 86400, 1)
+                time_ago_str = f"({days_ago:.1f} day{'s' if days_ago != 1 else ''} ago)"
+
+            latest_timestamp = f"{latest_timestamp_str} {time_ago_str}"
+
+        batch_stats[batch_name] = {
+            "unique_metrics": len(df["metric_name"].unique()),
+            "latest_timestamp": latest_timestamp,
+        }
+
+    main_content = Div(
+        Card(
+            DivLAligned(
+                H2(
+                    "Anomstack",
+                    P(
+                        "Painless open source anomaly detection for your metrics 📈📉🚀",
+                        cls=TextPresets.muted_sm,
+                    ),
+                    cls="mb-2",
+                ),
+                A(
+                    DivLAligned(UkIcon("github")),
+                    href="https://github.com/andrewm4894/anomstack",
+                    target="_blank",
+                    cls="uk-button uk-button-secondary",
+                    uk_tooltip="View on GitHub",
+                ),
+                style="justify-content: space-between;",
+                cls="mb-6",
+            ),
+            # Show warning if no metric batches
+            (
+                Div(
+                    DivLAligned(
+                        UkIcon("alert-triangle"),
+                        P(
+                            "No metric batches found. Is Dagster running?",
+                            cls=TextPresets.muted_sm,
+                        ),
+                        cls="space-x-2 p-4 bg-yellow-50 text-yellow-700 rounded-md",
+                    ),
+                    cls="mb-6",
+                )
+                if not app.state.metric_batches
+                else None
+            ),
+            (
+                Grid(
+                    *[
+                        Card(
+                            DivLAligned(
+                                Div(
+                                    H4(batch_name, cls="mb-2"),
+                                    DivLAligned(
+                                        Div(
+                                            DivLAligned(
+                                                UkIcon("activity", cls="text-blue-500"),
+                                                P(
+                                                    f"{batch_stats[batch_name]['unique_metrics']} metrics",
+                                                    cls=TextPresets.muted_sm,
+                                                ),
+                                                cls="space-x-2",
+                                            ),
+                                            DivLAligned(
+                                                UkIcon("clock", cls="text-green-500"),
+                                                P(
+                                                    f"{batch_stats[batch_name]['latest_timestamp']}",
+                                                    cls=TextPresets.muted_sm,
+                                                ),
+                                                cls="space-x-2",
+                                            ),
+                                            cls="space-y-2",
+                                        )
+                                    ),
+                                ),
+                                Button(
+                                    "View Metrics",
+                                    hx_get=f"/batch/{batch_name}",
+                                    hx_push_url=f"/batch/{batch_name}",
+                                    hx_target="#main-content",
+                                    hx_indicator="#loading",
+                                    cls=ButtonT.primary,
+                                ),
+                                style="justify-content: space-between;",
+                                cls="flex-row items-center",
+                            ),
+                            cls="p-6 hover:border-primary transition-colors duration-200",
+                        )
+                        for batch_name in app.state.metric_batches
+                    ],
+                    cols=3,
+                    gap=4,
+                )
+                if app.state.metric_batches
+                else None
+            ),
+            cls="p-6",
+        ),
+        id="main-content",
+    )
+
+    # For HTMX requests, return only the main content
+    if is_htmx:
+        return main_content
+
+    # For full page loads, return the complete layout
+    return (
+        Title("Anomstack"),
+        script,
+        Div(
+            Safe('<span class="htmx-indicator">Loading...</span>'),
+            id="loading",
+            cls="loading-indicator",
+        ),
+        main_content,
+    )
+
+
+@rt("/batch/{batch_name}")
+def get_batch_view(batch_name: str, session, initial_load: int = DEFAULT_LOAD_N_CHARTS):
+    # First, ensure we have the data and stats calculated
+    if batch_name not in app.state.df_cache:
+        app.state.df_cache[batch_name] = get_data(
+            app.state.specs_enabled[batch_name], max_n=DEFAULT_ALERT_MAX_N
+        )
+        app.state.calculate_metric_stats(batch_name)
+    elif batch_name not in app.state.stats_cache:
+        app.state.calculate_metric_stats(batch_name)
+
+    metric_stats = app.state.stats_cache[batch_name]
+    total_metrics = len(metric_stats)
+    remaining_metrics = total_metrics - initial_load
+
+    script = Script(
+        f"""
+        document.querySelectorAll('.top-nav li').forEach(li => {{
+            li.classList.remove('uk-active');
+            if (li.querySelector('a').textContent.trim() === '{batch_name}') {{
+                li.classList.add('uk-active');
+            }}
+        }});
+        window.scrollTo({{ top: 0, behavior: 'smooth' }});
+    """
+    )
+
+    # Update the load more button text
+    load_next = min(DEFAULT_LOAD_N_CHARTS, remaining_metrics)
+    return Div(
+        _create_controls(batch_name),
+        Div(
+            *[
+                ChartManager.create_chart_placeholder(
+                    stat["metric_name"], i, batch_name
+                )
+                for i, stat in enumerate(metric_stats[:initial_load])
+            ],
+            id="charts-container",
+            cls=f"grid grid-cols-{2 if app.state.two_columns else 1} gap-4",
+        ),
+        # Load More button with updated text
+        Div(
+            Button(
+                f"Load next {load_next} of {remaining_metrics}",
+                hx_get=f"/batch/{batch_name}/load-more/{initial_load}",
+                hx_target="#charts-container",
+                hx_swap="beforeend",
+                hx_indicator="#loading",
+                cls=ButtonT.secondary,
+                style="width: 100%; margin-top: 1rem;",
+                disabled=remaining_metrics <= 0,
+            ),
+            id="load-more-container",
+        ),
+        script,
+    )
+
+
+@rt("/batch/{batch_name}/chart/{chart_index}")
+def get(batch_name: str, chart_index: int):
+    df = app.state.df_cache[batch_name]
+    metric_stats = app.state.stats_cache[batch_name]
+    metric_name = metric_stats[chart_index]["metric_name"]
+    anomaly_rate = metric_stats[chart_index]["anomaly_rate"]
+    avg_score = metric_stats[chart_index]["avg_score"]
+
+    if batch_name not in app.state.chart_cache:
+        app.state.chart_cache[batch_name] = {}
+
+    if chart_index not in app.state.chart_cache[batch_name]:
+        df_metric = df[df["metric_name"] == metric_name]
+        fig = ChartManager.create_chart(df_metric, metric_name, chart_index)
+        app.state.chart_cache[batch_name][chart_index] = fig
+
+    return Card(
+        Style(
+            """
+            .uk-card-header {
+                padding: 1rem;  /* Consistent padding */
+            }
+            .uk-card-body {
+                padding: 1rem;  /* Consistent padding */
+            }
+        """
+        ),
+        Safe(app.state.chart_cache[batch_name][chart_index]),
+        header=Div(
+            H4(metric_name, cls="mb-1"),
+            DivLAligned(
+                P(
+                    f"Anomaly Rate: {anomaly_rate:.1%}",
+                    cls="text-sm text-muted-foreground",
+                ),
+                P(f"Avg Score: {avg_score:.1%}", cls="text-sm text-muted-foreground"),
+                style="gap: 1rem;",
+            ),
+        ),
+        id=f"chart-{chart_index}",
+        cls="mb-1",
+    )
+
+
+@rt("/batch/{batch_name}/refresh")
+def get(batch_name: str, session):
+    app.state.clear_batch_cache(batch_name)
+
+    return get_batch_view(batch_name, session, initial_load=DEFAULT_LOAD_N_CHARTS)
+
+
+@rt("/batch/{batch_name}/load-more/{start_index}")
+def get(batch_name: str, start_index: int):
+    metric_stats = app.state.stats_cache[batch_name]
+    remaining_metrics = len(metric_stats) - (start_index + DEFAULT_LOAD_N_CHARTS)
+    load_next = min(DEFAULT_LOAD_N_CHARTS, remaining_metrics)
+
+    return [
+        *[
+            ChartManager.create_chart_placeholder(stat["metric_name"], i, batch_name)
+            for i, stat in enumerate(
+                metric_stats[start_index : start_index + DEFAULT_LOAD_N_CHARTS],
+                start=start_index,
+            )
+        ],
+        # Update the load more button with new count and disabled state
+        Div(
+            Button(
+                (
+                    f"Load next {load_next} of {remaining_metrics}"
+                    if remaining_metrics > 0
+                    else "No more metrics"
+                ),
+                hx_get=f"/batch/{batch_name}/load-more/{start_index + DEFAULT_LOAD_N_CHARTS}",
+                hx_target="#charts-container",
+                hx_swap="beforeend",
+                hx_indicator="#loading",
+                cls=ButtonT.secondary,
+                style="width: 100%; margin-top: 1rem;",
+                disabled=remaining_metrics <= 0,
+            ),
+            id="load-more-container",
+            hx_swap_oob="true",
+        ),
+    ]
+
+
+@rt("/batch/{batch_name}/search")
+def post(batch_name: str, search: str = ""):
+    import re
+
+    metric_stats = app.state.stats_cache[batch_name]
+    try:
+        pattern = re.compile(search, re.IGNORECASE) if search else None
+        filtered_stats = [
+            stat
+            for stat in metric_stats
+            if not pattern or pattern.search(stat["metric_name"])
+        ]
+        placeholders = []
+        for i, stat in enumerate(filtered_stats):
+            placeholders.append(
+                ChartManager.create_chart_placeholder(
+                    stat["metric_name"], i, batch_name
+                )
+            )
+        return (
+            Div(*placeholders) if placeholders else Div(P("No matching metrics found"))
+        )
+    except re.error:
+        return Div(P("Invalid search pattern"))
+
+
+@rt("/batch/{batch_name}/update-n")
+def post(batch_name: str, alert_max_n: int = DEFAULT_ALERT_MAX_N, session=None):
+    app.state.alert_max_n[batch_name] = alert_max_n
+    app.state.clear_batch_cache(batch_name)
+
+    app.state.df_cache[batch_name] = get_data(
+        app.state.specs_enabled[batch_name], max_n=alert_max_n
+    )
+    app.state.calculate_metric_stats(batch_name)
+
+    # Return the full page content with proper URL update
+    return get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+
+
+@rt("/batch/{batch_name}/toggle-size")
+def post(batch_name: str, session=None):
+    app.state.small_charts = not app.state.small_charts
+    app.state.chart_cache.clear()  # Clear cache to regenerate charts
+    return get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+
+
+@rt("/batch/{batch_name}/toggle-theme")
+def post(batch_name: str, session=None):
+    app.state.dark_mode = not app.state.dark_mode
+    app.state.chart_cache.clear()  # Clear cache to regenerate charts
+
+    # Add script to toggle dark mode class on body
+    script = Script(
+        """
+        document.body.classList.toggle('dark-mode');
+    """
+    )
+
+    response = get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+    return Div(script, response)
+
+
+@rt("/batch/{batch_name}/toggle-columns")
+def post(batch_name: str, session=None):
+    app.state.two_columns = not app.state.two_columns
+    return get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+
+
+@rt("/batch/{batch_name}/toggle-markers")
+def post(batch_name: str, session=None):
+    app.state.show_markers = not app.state.show_markers
+    app.state.chart_cache.clear()  # Clear cache to regenerate charts
+    return get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+
+
+@rt("/batch/{batch_name}/toggle-legend")
+def post(batch_name: str, session=None):
+    app.state.show_legend = not app.state.show_legend
+    app.state.chart_cache.clear()  # Clear cache to regenerate charts
+    return get_batch_view(
+        batch_name, session=session, initial_load=DEFAULT_LOAD_N_CHARTS
+    )
+
+
+@rt("/batch/{batch_name}/update-line-width")
+def post(batch_name: str, line_width: int = 2):
+    try:
+        # Ensure line width is within valid range
+        app.state.line_width = max(1, min(10, int(line_width)))
+        app.state.chart_cache.clear()  # Clear cache to regenerate charts
+
+        # Only regenerate the charts, maintaining the id and grid layout
+        metric_stats = app.state.stats_cache[batch_name]
+        return [
+            *[
+                ChartManager.create_chart_placeholder(
+                    stat["metric_name"], i, batch_name
+                )
+                for i, stat in enumerate(metric_stats)
+            ]
+        ]
+    except ValueError:
+        return "Invalid line width value"
